+++ trading/trader/modes/__init__.py
@@ -1,0 +1,2 @@
+from .Default import *
+from .Advanced import *